add configurations manager
It’s now possible to have a few configurations and user can select any
one to use. User can edit the dospad.cfg, and also the ui.cfg which
defines gamepad layout and keybindings. Every configuration can have an
icon.png.

The configurations are saved under Documents/config folder. There will
be always a  “default” configuration. Every configuration has its own
folder.

Require application restart to take effect.

update app icon.

+++ dosbox/src/shell/shell.cpp
@@ -386,7 +386,7 @@ void DOS_Shell::Run(void) {
                 if (fp != NULL)
                 {
                     char buf[256];
-                    int sectionLength = strlen(dospad_launch_section);
+                    size_t sectionLength = strlen(dospad_launch_section);
                     while (fgets(buf, 256, fp))
                     {
                         if (strncmp(buf, dospad_launch_section, sectionLength) == 0)
